+++ src/SNGBE/PaymentException.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Igor Bodnar <bodnar_i@mail.ru>
+ * Date: 03.08.2015
+ * Time: 9:56
+ */
+
+namespace SNGBE;
+
+/**
+ * Class PaymentException
+ *
+ * @package SNGBE
+ */
+class PaymentException extends \Exception
+{
+}