+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+0.2.0
+-----
+Big change of architecture:
+
+* Payment now creates by SNGBEcomm::createPayment()
+* Remove class Error in flavour of Payment::isError()
+* Remove all static methods
+* Use PHPUnit instead of vierbergenlars/simpletest
+* Move all classes to SNGBEcomm namespace
+* Use PSR-0 for autoloading classes
+
 0.1.3
 -----
 * Create basic payment behavior ([@zatsepin](mailto:yu.zatsepin@sngb.ru))